requirements.txt
streamlit
mysql-connector-python
pandas
numpy
bcrypt
mysql-connector-python